extras/pyproject.toml
[project]
name = "zoho-crm-data-population"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "zohocrmsdk2-1>=3.0.0",
    "zohocrmsdk2-python",
]
